OAK-1717 Concurrent updates of ordered index in cluster may fail
made long the MAX for the count to allow better definitions.

+++ oak-core/src/main/java/org/apache/jackrabbit/oak/plugins/index/property/strategy/IndexStoreStrategy.java
@@ -71,6 +71,6 @@ public interface IndexStoreStrategy {
          * @param max
          * @return
          */
-        long count(NodeState indexMeta, Filter.PropertyRestriction pr, int max);
+        long count(NodeState indexMeta, Filter.PropertyRestriction pr, long max);
     }
 }
